refactor(products): split list into client component, centralize analysis types
Extract the products page into products-list-client.tsx with a loading.tsx
skeleton; move shared analysis types (Verdict, scores, third-party recipients,
document summaries, crawl errors) into types/analysis.ts and re-export via the
barrel; retype the overview components, sources-list, and pipeline routes.
