+++ packages/frontend/components/dashboard/overview/sharing-map.tsx
@@ -11,13 +11,7 @@ import {
 import { Badge } from "@/components/ui/badge";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { cn } from "@/lib/utils";
-
-interface ThirdPartyRecipient {
-  recipient: string;
-  data_shared: string[];
-  purpose?: string | null;
-  risk_level: "low" | "medium" | "high";
-}
+import type { ThirdPartyRecipient } from "@/types";
 
 interface SharingMapProps {
   thirdPartyDetails?: ThirdPartyRecipient[] | null;
