+++ packages/frontend/types/index.ts
@@ -1,1 +1,20 @@
 export type { Product } from "./product";
+export type {
+  Verdict,
+  DataPurposeLink,
+  ThirdPartyRecipient,
+  DetailedScore,
+  DetailedScores,
+  PrivacySignalsData,
+  CoverageItem,
+  ComplianceBreakdown,
+  ProductOverview,
+  EvidenceSpan,
+  KeypointWithEvidence,
+  CriticalClause,
+  DocumentRiskBreakdown,
+  DocumentSection,
+  DocumentSummary,
+  CrawlError,
+  FailedCrawlJob,
+} from "./analysis";
